requirements.txt
openai-whisper
pyannote.audio
torch
pydub
moviepy
requests
PyGithub
python-dateutil
